+++ middlewares/userAuthValidator.js
@@ -1,20 +1,27 @@
 const { body, validationResult } = require('express-validator');
 
 const userAuthValidation = [
-body('username').isEmail(),
+    body('username').isEmail(),
 
-body('password').isLength({ min: 5 }),
+    body('password').isLength({ min: 5 }),
 
 ];
 
-const validationHandler = (req,res,next) => {
+const registerValidation = [
+    ...userAuthValidation,
+    //validate name and last name
+    body('firstName').isString(),
+    body('lastName').isString(),
+]
+
+const validationHandler = (req, res, next) => {
     const errors = validationResult(req);
-    if(errors.isEmpty()){
+    if (errors.isEmpty()) {
         return next();
     }
-        return res.status(400).json({
-            errors: errors.array()
-        });
+    return res.status(400).json({
+        errors: errors.array()
+    });
 };
 
-module.exports = { userAuthValidation, validationHandler };
+module.exports = { userAuthValidation , registerValidation, validationHandler };
